new bouncing balls script and put all scripts into their own files to reduce clutter
